requirements.txt
Flask
gunicorn
psycopg2-binary
qrcode
werkzeug
qrcode
Pillow